refactor: 역할 별 컴포넌트 분리 및 당장 분리할 수 있는 함수 분리
